% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_ASN}
\alias{calc_ASN}
\title{Helper function to calculate the average sample size}
\usage{
calc_ASN(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate the average sample size
}
