% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_ASNP}
\alias{calc_ASNP}
\title{Calculate the average placebo group sample size}
\usage{
calc_ASNP(P, D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Calculate the average placebo group sample size
}
