% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_c}
\alias{calc_c}
\title{Helper function to calculate allocation ratios from sample sizes}
\usage{
calc_c(nT1, D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate allocation ratios from sample sizes
}
