% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_probability_functions.R
\name{calc_conditional_local_rejection_probs}
\alias{calc_conditional_local_rejection_probs}
\title{Calculate the (local) conditional type I errors of both hypothesis given both interim test statistics.}
\usage{
calc_conditional_local_rejection_probs(Z_TP1, Z_TC1, D, mu_vec = D$mu_vec$H0)
}
\arguments{
\item{Z_TP1}{First stage Z-test statistic of the superiority test of treatment over placebo.}

\item{Z_TC1}{First stage Z-test statistic of the non-inferiority test of treatment to the active control.}

\item{D}{list containing all parameters of a the trial design.}

\item{mu_vec}{vector of means of the test statistic c(Z_TP1, Z_TP2, Z_TC1, Z_TC1).}
}
\value{
named numeric vector with both conditional type I errors.
}
\description{
Calculate the (local) conditional type I errors of both hypothesis given both interim test statistics.
}
