% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_probability_functions.R
\name{calc_conditional_power}
\alias{calc_conditional_power}
\title{Calculate the conditional power to reject both hypothesis given both interim test statistics.}
\usage{
calc_conditional_power(Z_TP1, Z_TC1, D)
}
\arguments{
\item{Z_TP1}{First stage Z-test statistic of the superiority test of treatment over placebo.}

\item{Z_TC1}{First stage Z-test statistic of the non-inferiority test of treatment to the active control.}

\item{D}{list containing all parameters of a the trial design.}
}
\value{
numeric value of the conditional power.
}
\description{
Calculate the conditional power to reject both hypothesis given both interim test statistics.
}
