% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_cumc}
\alias{calc_cumc}
\title{Helper function to calculate "cumulative allocation ratio" from design object}
\usage{
calc_cumc(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate "cumulative allocation ratio" from design object
}
