% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_cumn}
\alias{calc_cumn}
\title{Helper function to calculate cumulative sample size after second stage from design object}
\usage{
calc_cumn(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate cumulative sample size after second stage from design object
}
