% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_final_state_probs}
\alias{calc_final_state_probs}
\title{Helper function to calculate the final state probabilities}
\usage{
calc_final_state_probs(hypothesis = "H0", D)
}
\arguments{
\item{hypothesis}{string describing the hypothesis, either "H0" or "H1".}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate the final state probabilities
}
