% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_gamma}
\alias{calc_gamma}
\title{Helper function to calculate gamma factors from group variances and cumulative allocation ratios}
\usage{
calc_gamma(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate gamma factors from group variances and cumulative allocation ratios
}
