% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_local_alphas}
\alias{calc_local_alphas}
\title{Helper function to calculate the local type I error rates of a Design}
\usage{
calc_local_alphas(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate the local type I error rates of a Design
}
