% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_local_rejection_boundaries}
\alias{calc_local_rejection_boundaries}
\title{Helper function to calculate the local rejection boundaries of group sequential testing
procedure associated with the hypothesis belong to the groups argument}
\usage{
calc_local_rejection_boundaries(groups = "TP", D)
}
\arguments{
\item{groups}{string describing the pair of groups. Either "TP" or "TC".}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate the local rejection boundaries of group sequential testing
procedure associated with the hypothesis belong to the groups argument
}
