% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_mu_vec}
\alias{calc_mu_vec}
\title{Helper function to calculate expected value from design object}
\usage{
calc_mu_vec(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate expected value from design object
}
