% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_mu_wo_nT1}
\alias{calc_mu_wo_nT1}
\title{Helper function to calculate expected value for n_{1,T}=1}
\usage{
calc_mu_wo_nT1(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
This function exists for performance optimization reasons, as it saves some calculations in the long run.
}
