% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_n_from_c}
\alias{calc_n_from_c}
\title{Helper function to calculate other n's given n_{1,T} and allocation ratios from design object}
\usage{
calc_n_from_c(nT1, D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate other n's given n_{1,T} and allocation ratios from design object
}
