% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_helper_functions_nonsequential_futility.R
\name{calc_nT1_wrt_bTC2e}
\alias{calc_nT1_wrt_bTC2e}
\title{Helper function to calculate the required sample size (of the stage 1 treatment group)
to achieve the target power given the b2TCe}
\usage{
calc_nT1_wrt_bTC2e(b2TCe, D)
}
\arguments{
\item{b2TCe}{second stage critical value for the TC testing problem.}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
This is required in designs with nonsequential futility testing, as choosing locally valid designs
is insufficient to gurantee type I error control.
}
