% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_prob_reject_both}
\alias{calc_prob_reject_both}
\title{Calculate the probability to reject both hypotheses given a choice of alternative or null point hypothesis and a design object}
\usage{
calc_prob_reject_both(hypothesis = "H1", D = list())
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Calculate the probability to reject both hypotheses given a choice of alternative or null point hypothesis and a design object
}
