% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_prob_reject_both_singlestage}
\alias{calc_prob_reject_both_singlestage}
\title{Helper function to calculate the probability to reject both hypotheses
given the mean of the normal test statistic vector c(Z_TP1, Z_TC1).}
\usage{
calc_prob_reject_both_singlestage(mu_vec, D)
}
\arguments{
\item{mu_vec}{vector of means of the test statistic c(Z_TP1, Z_TP2, Z_TC1, Z_TC1).}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate the probability to reject both hypotheses
given the mean of the normal test statistic vector c(Z_TP1, Z_TC1).
}
