% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_prob_reject_both_wrt_mu_vec}
\alias{calc_prob_reject_both_wrt_mu_vec}
\title{Calculate the probability to reject both hypothesis given an expected value for the test statistic and a design object}
\usage{
calc_prob_reject_both_wrt_mu_vec(mu_vec, D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Calculate the probability to reject both hypothesis given an expected value for the test statistic and a design object
}
