% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_rho}
\alias{calc_rho}
\title{Helper function to calculate rho from design object}
\usage{
calc_rho(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate rho from design object
}
