% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_Sigma}
\alias{calc_Sigma}
\title{Helper function to calculate the covariance matrix from the group variances, cumulative allocation ratios
and gamma factors}
\usage{
calc_Sigma(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to calculate the covariance matrix from the group variances, cumulative allocation ratios
and gamma factors
}
