% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_worst_type_I_error}
\alias{calc_worst_type_I_error}
\title{Helper function to calculate the maximal probability of rejecting the non-inferiority hypothesis
in the testing procedure featuring nonsequential futility, given a point hypothesis for
the superiority hypothesis.}
\usage{
calc_worst_type_I_error(bTC2e, D)
}
\arguments{
\item{bTC2e}{second stage critical value for the TC testing problem.}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
This is required in designs with nonsequential futility testing, as choosing locally valid designs
is insufficient to gurantee type I error control.
}
