% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{create_Design}
\alias{create_Design}
\title{Helper function to create a design object}
\usage{
create_Design(
  type_I_error = 0.025,
  type_II_error = 0.1,
  Delta = 0.3,
  alternative_TP = 0.6,
  lambda = 1,
  kappa = 0,
  sigma_T = 1,
  sigma_P = 1,
  sigma_C = 1,
  tol = 1e-07,
  maxpts = 4097,
  maxeval = 150
)
}
\arguments{
\item{maxeval}{}
}
\value{

}
\description{
Helper function to create a design object
}
