% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{finalStateProb}
\alias{finalStateProb}
\title{Function to calculate the final state probabilities}
\usage{
finalStateProb(hypothesis = "H0", D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Function to calculate the final state probabilities
}
