% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_methods.R
\name{init_successive_optimum_designs}
\alias{init_successive_optimum_designs}
\title{Optimizes designs 1 through 6 in succession, using the previous optimum as the starting point for the next design.}
\usage{
init_successive_optimum_designs(
  boundaryvals = list(),
  ncores = NULL,
  designs = 1:6
)
}
\arguments{
\item{boundaryvals}{list of boundary conditions for the designs to be optimized}

\item{ncores}{number of cores to use for parallelization}

\item{designs}{vector of designs containing the numbers 1 through 6}
}
\value{

}
\description{
Optimizes designs 1 through 6 in succession, using the previous optimum as the starting point for the next design.
}
