% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_probability_functions.R
\name{meandiff_to_Z}
\alias{meandiff_to_Z}
\title{Convenience function for conversion of differences in means to Z statistic}
\usage{
meandiff_to_Z(meandiff_TP, meandiff_TC, gamma, nT1)
}
\arguments{
\item{meandiff_TP}{numeric}

\item{meandiff_TC}{numeric}

\item{gamma}{list containing the "gamma" factors.
gamma[[1]][["TP"]] = sqrt(var[["T"]] / cumc[["T"]] + var[["T"]] / cumc[["P"]]), etc.}

\item{nT1}{first stage sample size of the treatment group.}
}
\value{
Z statistic Vector
}
\description{
Calculates Z statistic vector from differences in means, gamma list and number of patients in
the treatment group in stage 1
}
