% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{objective_closed_testing}
\alias{objective_closed_testing}
\title{Design 6}
\usage{
objective_closed_testing(x, D, return_everything = FALSE)
}
\arguments{
\item{return_everything}{}
}
\value{

}
\description{
Design 6
}
