% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{objective_fully_sequential}
\alias{objective_fully_sequential}
\title{Design 5}
\usage{
objective_fully_sequential(x, D, return_everything = FALSE)
}
\arguments{
\item{return_everything}{}
}
\value{

}
\description{
Design 5
}
