% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{objective_no_futility}
\alias{objective_no_futility}
\title{Design 3}
\usage{
objective_no_futility(x, D, return_everything = FALSE)
}
\arguments{
\item{return_everything}{}
}
\value{

}
\description{
Design 3
}
