% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{objective_no_futility_fixed_c}
\alias{objective_no_futility_fixed_c}
\title{Design 2}
\usage{
objective_no_futility_fixed_c(x, D, return_everything = FALSE)
}
\arguments{
\item{return_everything}{}
}
\value{

}
\description{
Design 2
}
