% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{objective_nonbinding_futility}
\alias{objective_nonbinding_futility}
\title{Design 4}
\usage{
objective_nonbinding_futility(x, D, return_everything = FALSE)
}
\arguments{
\item{return_everything}{}
}
\value{

}
\description{
Design 4
}
