% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{objective_onestage}
\alias{objective_onestage}
\title{Objective function for single-stage gold-standard designs}
\usage{
objective_onestage(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Objective function for single-stage gold-standard designs
}
