% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{objective_single_stage}
\alias{objective_single_stage}
\title{Design 1}
\usage{
objective_single_stage(x, D, return_everything = FALSE)
}
\arguments{
\item{return_everything}{}
}
\value{

}
\description{
Design 1
}
