% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{objective_singlestage}
\alias{objective_singlestage}
\title{Objective function for single-stage designs}
\usage{
objective_singlestage(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Objective function for single-stage designs
}
