% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_functions.R
\name{objective_twostage}
\alias{objective_twostage}
\title{Objective function for two-stage gold-standard designs}
\usage{
objective_twostage(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Objective function for two-stage gold-standard designs
}
