% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_optimization_methods.R
\name{opt_objective_fully_sequential}
\alias{opt_objective_fully_sequential}
\title{Optimal design parameters for design 5}
\usage{
opt_objective_fully_sequential(start = NULL, D = create_Design(maxeval = 10))
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Optimal design parameters for design 5
}
