% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_methods.R
\name{opt_objective_no_futility}
\alias{opt_objective_no_futility}
\title{Optimal design parameters for design 3}
\usage{
opt_objective_no_futility(start = NULL, D = create_Design(maxeval = 10))
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Optimal design parameters for design 3
}
