% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_methods.R
\name{opt_objective_no_futility_fixed_c}
\alias{opt_objective_no_futility_fixed_c}
\title{Optimal design parameters for design 2}
\usage{
opt_objective_no_futility_fixed_c(start = NULL, D = create_Design())
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Optimal design parameters for design 2
}
