% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_optimization_methods.R
\name{opt_objective_single_stage}
\alias{opt_objective_single_stage}
\title{Optimal design parameters for design 1}
\usage{
opt_objective_single_stage(start = NULL, D = create_Design())
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Optimal design parameters for design 1
}
