% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalGoldstandardDesigns.R
\docType{package}
\name{OptimalGoldstandardDesigns}
\alias{OptimalGoldstandardDesigns}
\title{OptimalGoldstandardDesigns}
\description{
Methods to calculate optimal design parameters
for one- and two-stage three-arm group-sequential gold-standard
non-inferiority trial designs
with or without binding or nonbinding futility boundaries,
as described in \insertCite{meis2023optimization}{OptimalGoldstandardDesigns}.
}
\references{
\insertAllCited{}
}
