% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{padd_whitespace}
\alias{padd_whitespace}
\title{Add whitespace padding to string}
\usage{
padd_whitespace(str)
}
\arguments{
\item{str}{a character string}
}
\value{
string with whitespace padding until the full console width
}
\description{
Add whitespace padding to string
}
