% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmv_upper_smaller_slower_fix.R
\name{pmvt_}
\alias{pmvt_}
\title{mvtnorm::pmvt, but returns 0 if any lower boundary is larger than
any upper boundary}
\usage{
pmvt_(upper, lower, ...)
}
\arguments{
\item{upper}{the vector of upper limits of length n.}

\item{lower}{the vector of lower limits of length n.}

\item{...}{additional parameters passed to \code{\link[mvtnorm]{pmvt}}.}
}
\value{
The evaluated distribution function is returned, if \code{keepAttr} is true, with attributes
\item{error}{estimated absolute error}
\item{msg}{status message(s).}
\item{algorithm}{a \code{\link{character}} string with \code{class(algorithm)}.}
}
\description{
mvtnorm::pmvt, but returns 0 if any lower boundary is larger than
any upper boundary
}
\seealso{
\code{\link[mvtnorm]{pmvt}}
}
