% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{print.GoldStandardDesign}
\alias{print.GoldStandardDesign}
\title{Title}
\usage{
\method{print}{GoldStandardDesign}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{

}
\description{
Title
}
