% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{print.OneStageGoldStandardDesign}
\alias{print.OneStageGoldStandardDesign}
\title{Printing method for optimal single-stage goldstandard designs}
\usage{
\method{print}{OneStageGoldStandardDesign}(x, ...)
}
\arguments{
\item{x}{An object of class OneStageGoldStandardDesign}

\item{...}{additional parameters}
}
\value{

}
\description{
Printing method for optimal single-stage goldstandard designs
}
\examples{
1 + 1
}
