% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_ASNP}
\alias{calc_ASNP}
\title{Calculate the average placebo group sample size}
\usage{
calc_ASNP(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Calculate the average placebo group sample size
}
