% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_c}
\alias{calc_c}
\title{Helper function to calculate allocation ratios from stagewise sample sizes}
\usage{
calc_c(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate allocation ratios from stagewise sample sizes
}
