% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_cumc}
\alias{calc_cumc}
\title{Helper function to calculate "cumulative allocation ratio" from stagewise allocation ratios}
\usage{
calc_cumc(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate "cumulative allocation ratio" from stagewise allocation ratios
}
