% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_cumn}
\alias{calc_cumn}
\title{Helper function to calculate cumulative sample sizes from stagewise sample sizes}
\usage{
calc_cumn(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate cumulative sample sizes from stagewise sample sizes
}
