% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_mu_vec}
\alias{calc_mu_vec}
\title{Helper function to calculate expected value of normal test statistic vector c(Z_TP1, Z_TP2, Z_TC1, Z_TP2)
under the null and alternative hypothesis given nT1, gamma and mu.}
\usage{
calc_mu_vec(D)
}
\arguments{
\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate expected value of normal test statistic vector c(Z_TP1, Z_TP2, Z_TC1, Z_TP2)
under the null and alternative hypothesis given nT1, gamma and mu.
}
