% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_n_from_c}
\alias{calc_n_from_c}
\title{Helper function to calculate other n's given n_{1,T} and allocation ratios}
\usage{
calc_n_from_c(nT1, D)
}
\arguments{
\item{nT1}{first stage sample size of the treatment group.}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate other n's given n_{1,T} and allocation ratios
}
