% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_helper_functions.R
\name{calc_nT1_wrt_bTC2e}
\alias{calc_nT1_wrt_bTC2e}
\title{Helper function to calculate the required sample size (of the stage 1 treatment group)
to achieve the target power given the bTC2e}
\usage{
calc_nT1_wrt_bTC2e(bTC2e, D)
}
\arguments{
\item{bTC2e}{second stage critical value for the TC testing problem.}

\item{D}{list containing all parameters of a the trial design.}
}
\description{
Helper function to calculate the required sample size (of the stage 1 treatment group)
to achieve the target power given the bTC2e
}
