% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_design_functions.R
\name{calc_Sigma}
\alias{calc_Sigma}
\title{Helper function to covariance matrix Sigma from design object}
\usage{
calc_Sigma(D)
}
\arguments{
\item{D}{}
}
\value{

}
\description{
Helper function to covariance matrix Sigma from design object
}
