% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalGoldstandardDesigns-package.R
\docType{package}
\name{OptimalGoldstandardDesigns}
\alias{OptimalGoldstandardDesigns}
\alias{OptimalGoldstandardDesigns-package}
\title{OptimalGoldstandardDesigns}
\description{
Methods to calculate optimal design parameters
for one- and two-stage three-arm group-sequential gold-standard
non-inferiority trial designs
with or without binding or nonbinding futility boundaries,
as described in \insertCite{meis2023optimization}{OptimalGoldstandardDesigns}.
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jan-imbi/OptimalGoldstandardDesigns}
}

}
\author{
\strong{Maintainer}: Jan Meis \email{meis@imbi.uni-heidelberg.de} (\href{https://orcid.org/0000-0001-5407-7220}{ORCID})

}
