% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_methods.R
\name{optimize_design_onestage}
\alias{optimize_design_onestage}
\title{Title}
\usage{
optimize_design_onestage(
  cP1 = 0.25,
  cC1 = 1,
  alpha = 0.025,
  beta = 0.2,
  alternative_TP = 0.4,
  alternative_TC = 0,
  Delta = 0.2,
  varT = 1,
  varP = 1,
  varC = 1,
  round_n = TRUE,
  kappa = 0,
  objective = quote(sum(unlist(n)) + kappa * n[[1]][["P"]]),
  inner_tol_objective = 1e-07,
  mvnorm_algorithm = mvtnorm::Miwa(steps = 500, checkCorr = FALSE, maxval = 1000),
  nloptr_x0 = NULL,
  nloptr_lb = NULL,
  nloptr_ub = NULL,
  nloptr_opts = list(algorithm = "NLOPT_LN_SBPLX", xtol_rel = 0.1, maxeval = 500),
  print_progress = TRUE,
  ...
)
}
\arguments{
\item{nloptr_opts}{}
}
\value{

}
\description{
Title
}
\examples{
1 + 1

}
