% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_methods.R
\name{optimize_design_twostage}
\alias{optimize_design_twostage}
\title{Title}
\usage{
optimize_design_twostage(
  cT2 = 1,
  cP1 = 0.25,
  cP2 = 0.25,
  cC1 = 1,
  cC2 = 1,
  bTP1f = 0,
  bTP1e = 2.3,
  bTC1f = 0,
  bTC1e = 2.3,
  alpha = 0.05,
  beta = 0.2,
  alternative_TP = 0.4,
  alternative_TC = 0,
  Delta = 0.2,
  varT = 1,
  varP = 1,
  varC = 1,
  nonbinding_futility = TRUE,
  always_both_futility_tests = TRUE,
  round_n = TRUE,
  lambda = 1,
  kappa = 0,
  nu = 0,
  objective = quote(sqrt(lambda)^2 * ASN[["H11"]] + (1 - sqrt(lambda)) * sqrt(lambda) *
    ASN[["H10"]] + (1 - sqrt(lambda)) * sqrt(lambda) * ASN[["H01"]] + (1 -
    sqrt(lambda))^2 * ASN[["H00"]] + kappa * (sqrt(lambda)^2 * ASNP[["H11"]] + (1 -
    sqrt(lambda)) * sqrt(lambda) * ASNP[["H10"]] + (1 - sqrt(lambda)) * sqrt(lambda) *
    ASNP[["H01"]] + (1 - sqrt(lambda))^2 * ASNP[["H00"]] + nu * cumn[[2]][["P"]]) + nu *
    (cumn[[2]][["T"]] + cumn[[2]][["P"]] + cumn[[2]][["C"]])),
  inner_tol_objective = 1e-07,
  mvnorm_algorithm = mvtnorm::Miwa(steps = 500, checkCorr = FALSE, maxval = 1000),
  nloptr_x0 = NULL,
  nloptr_lb = NULL,
  nloptr_ub = NULL,
  nloptr_opts = list(algorithm = "NLOPT_LN_SBPLX", xtol_rel = 0.1, maxeval = 500),
  ...
)
}
\arguments{
\item{cT2}{}

\item{cP1}{}

\item{cP2}{}

\item{cC1}{}

\item{cC2}{}

\item{bTP1f}{}

\item{bTP1e}{}

\item{bTC1e}{}

\item{alpha}{}

\item{beta}{}

\item{alternative_TP}{}

\item{alternative_TC}{}

\item{Delta}{}

\item{varT}{}

\item{varP}{}

\item{varC}{}

\item{nonbinding_futility}{}

\item{always_both_futility_tests}{}

\item{round_n}{}

\item{objective}{}

\item{inner_tol_objective}{}

\item{mvnorm_algorithm}{}

\item{nloptr_x0}{}

\item{nloptr_lb}{}

\item{nloptr_ub}{}

\item{nloptr_opts}{}

\item{i}{}
}
\value{

}
\description{
Title
}
\examples{
1 + 1

}
