% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{print.OneStageGoldStandardDesign}
\alias{print.OneStageGoldStandardDesign}
\title{Title}
\usage{
\method{print}{OneStageGoldStandardDesign}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
