% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{print.TwoStageGoldStandardDesign}
\alias{print.TwoStageGoldStandardDesign}
\title{Printing method for optimal two-stage goldstandard designs}
\usage{
\method{print}{TwoStageGoldStandardDesign}(x, ...)
}
\arguments{
\item{x}{An object of class TwoStageGoldStandardDesign}

\item{...}{additional parameters}
}
\value{

}
\description{
Printing method for optimal two-stage goldstandard designs
}
\examples{
1 + 1
}
