% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{print.TwoStageGoldStandardDesign}
\alias{print.TwoStageGoldStandardDesign}
\title{Printing method for optimal two-stage goldstandard designs}
\usage{
\method{print}{TwoStageGoldStandardDesign}(x, ...)
}
\arguments{
\item{x}{An object of class TwoStageGoldStandardDesign}

\item{...}{additional parameters}
}
\value{
returns the input x invisibly. This functions is used for its side effects, i.e. printing
design characteristics to the screen.
}
\description{
Printing method for optimal two-stage goldstandard designs
}
\examples{
D <- optimize_design_twostage(nloptr_opts = list(maxeval = 1, algorithm = "NLOPT_LN_SBPLX"))
D
}
