% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_display_functions.R
\name{print.TwoStageGoldStandardDesign}
\alias{print.TwoStageGoldStandardDesign}
\title{Printing method for optimal two-stage goldstandard designs}
\usage{
\method{print}{TwoStageGoldStandardDesign}(x, ...)
}
\arguments{
\item{x}{An object of class TwoStageGoldStandardDesign}

\item{...}{additional parameters}
}
\value{
returns the input x invisibly. This functions is used for its side effects, i.e. printing
design characteristics to the screen.
}
\description{
Printing method for optimal two-stage goldstandard designs
}
\examples{
# Should take about 15 seconds.
\donttest{
optimize_design_twostage(
  cT2 = 1,
  cP2 = quote(cP1),
  cC2 = quote(cC1),
  bTP1f = -Inf,
  bTC1f = -Inf,
  beta = 0.2,
  alternative_TP = 0.4,
  alternative_TC = 0,
  Delta = 0.2,
  binding_futility = TRUE,
  lambda = .9,
  kappa = 1,
  nloptr_opts = list(algorithm = "NLOPT_LN_SBPLX", ftol_rel = 1e-01)
)
}
}
